+++ SCrawler/My Project/AssemblyInfo.vb
@@ -32,6 +32,6 @@ Imports System.Runtime.InteropServices
 ' by using the '*' as shown below:
 ' <Assembly: AssemblyVersion("1.0.*")>
 
-<Assembly: AssemblyVersion("2024.4.26.0")>
-<Assembly: AssemblyFileVersion("2024.4.26.0")>
+<Assembly: AssemblyVersion("2024.5.4.0")>
+<Assembly: AssemblyFileVersion("2024.5.4.0")>
 <Assembly: NeutralResourcesLanguage("en")>